refactor(ip): adapt IP layer for ZMQ-based IPC integration

+++ src/ip/bit_utils.cpp
@@ -5,7 +5,7 @@ std::vector<uint8_t> byte_to_bits(uint8_t *bytes, IP &ip)
     std::vector<uint8_t> bits;
     ssize_t n = ip.nbytes.load();
     bits.reserve(n * 8);
-    for (size_t i = 0; i < n; ++i)
+    for (size_t i = 0; (ssize_t)i < n; ++i)
     {
         for (int j = 7; j >= 0; --j)
         {
